fix bug

+++ datasets/base_dataset.py
@@ -53,7 +53,7 @@ class BaseDataset(Dataset):
         self.labels = list(self.data_dict.keys())
 
         if self.classification:
-            self.label_list = F.one_hot(torch.tensor(self.label_list, dtype=torch.int))
+            self.label_list = F.one_hot(torch.tensor(self.label_list, dtype=torch.int64))
 
         return
